Move theme panel into a src folder
