+++ theme-panel/index.ts
@@ -1,0 +1,1 @@
+export { default as ThemePanel } from './src/ThemePanel';
