+++ theme-panel/src/ThemePanel.tsx
@@ -1,8 +1,8 @@
 import React, { FC, ReactElement } from 'react';
-import { useTheme } from '../theme-provider';
+import { useTheme } from '../../theme-provider';
 import { ActiveTab, Tab, TabBar, ThemePanelContainer } from './ThemePanel.styles';
-import { themeList } from '../theme-list';
-import { ThemeBase } from '../theme';
+import { themeList } from '../../theme-list';
+import { ThemeBase } from '../../theme';
 
 const ThemePanel: FC = ({ children }): ReactElement => {
   const { themeBase: currentThemeBase, setTheme } = useTheme();
