+++ theme/doc/components/colors-showcase/ColorsShowcase.tsx
@@ -3,7 +3,7 @@ import React, { ReactElement } from 'react';
 import ColorSample from '../color-sample/ColorSample';
 import { GroupLabel, Samples } from './ColorsShowcase.styles';
 import { useTheme } from '../../../../theme-provider';
-import ThemePanel from '../../../../theme-panel/ThemePanel';
+import { ThemePanel } from '../../../../theme-panel';
 
 const ColorsShowcase = (): ReactElement => {
   const { theme } = useTheme();
